refactor(ai): extract pass-1 tool selection into handler/pass1-tools.ts (U2)
